point 6 (in progress): the game boots - pilot select renders
All 133 remaining gameplay sources join the build (ship/ai, weapon,
mission, hud, menuui, missionui, ui, freespace2, ...). freespace.cpp's
WinMain is now main() (SDL2/POSIX); CD checks, launcher plumbing and
registry renderer selection neutered; software renderer is the mode.
Two-tier generated stubs: network/multi and sound families are silent
no-ops (OpenAL comes next), everything else traps loudly.

Runtime bring-up fixes: bm_set_components/color guns initialized for
software mode (1555 ARGB, transplanted from the Glide init); debris
vclip names were char* into string literals (strcpy target); duplicate
XSTR declaration in localize.h split the symbol; MSVC linker tolerated
duplicate globals (Buttons/Goal_text staticized); retail's commented-
out software-path palette code in playermenu restored, pointed at the
background PCX palette (ChoosePilotPalette isn't in the retail VPs).

The binary boots from rundir against the GOG VPs: full table parse
(ships/weapons/sounds), gamesequence runs, and the pilot-select
screen renders text and UI through the ported font/blit stack.
Temporary FS2_FRAME_DUMP hook in grsoft for headless verification.

+++ code/missionui/missionbrief.cpp
@@ -716,9 +716,9 @@ void brief_buttons_init()
 //	brief_get_closeup_icon()
 //
 //
-uint brief_get_closeup_icon()
+uintptr_t brief_get_closeup_icon()
 {
-	return (uint)Closeup_icon;
+	return (uintptr_t)Closeup_icon;
 }
 
 // stop showing the closeup view of an icon
